requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.5.1
    # via django
dj-database-url==0.5.0
    # via -r requirements.in
django==4.0.4
    # via
    #   -r requirements.in
    #   django-authtools
    #   djangorestframework
django-authtools @ https://github.com/fusionbox/django-authtools/archive/authtools-2.0.zip
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
pytz==2022.1
    # via djangorestframework
sqlparse==0.4.2
    # via django
